% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRangesFromUniprot.R
\name{CollapseSpanRanges}
\alias{CollapseSpanRanges}
\title{CollapseSpanRanges Handle a set of range to have only one range per position. This will allow us to use the}
\usage{
CollapseSpanRanges(SpanRanges)
}
\arguments{
\item{SpanRanges}{IRanges that will be used for HMTL formatting}
}
\value{
IRanges with class and title information for HTML span tag
}
\description{
CollapseSpanRanges Handle a set of range to have only one range per position. This will allow us to use the
}
\examples{
library(GenomicRanges)

VariationRanges<-GetRangesFromUniprot("O94875-10")
VariationRanges
DuplicatedVariationRanges<-VariationRanges
mcols(DuplicatedVariationRanges)$Title<-paste("Duplicated ",mcols(DuplicatedVariationRanges)$Title)
VariationRanges<-c(VariationRanges,DuplicatedVariationRanges)
Collapsed<-CollapseSpanRanges(VariationRanges)
VariationRanges
Collapsed
}
