% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareTwoSequences.R
\name{CompareTwoSequences}
\alias{CompareTwoSequences}
\title{CompareTwoSequences Compare two AA sequences in a blast results}
\usage{
CompareTwoSequences(QuerySequence, SubjectSequence, QueryStart = 12)
}
\arguments{
\item{QuerySequence}{The query sequence in the blast result file, the query sequence correspond to the sequence of the protein bank aligned to the knowned protein bank (Uniprot for example).}

\item{SubjectSequence}{The subject sequence in the blast results file, the subject sequence correspond to the known protein (typically one of UNIPROT) on which the query have been aligned.}

\item{QueryStart}{In the case where the matching part of the query do not beggin with its N-Term extremity.}
}
\value{
IRanges with metadata corresponding to the differences between the two sequences, SAAV, Insertion, Deletion
}
\description{
CompareTwoSequences Compare two AA sequences in a blast results
}
\examples{
R_FilePath <- system.file("extdata",
 "blast_rescue.example.txt",
  package = "isoformnspectRe")
Rescue<-read.table(
 R_FilePath,
 header=TRUE,
 stringsAsFactors = FALSE)
CompareTwoSequences(
 as.vector(Rescue$QuerySequence[1]),
 as.vector(Rescue$SubjectSequence[1]),
 as.vector(Rescue$QueryStart[1]))
}
