% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATAFRAME_MakeALineForAIsoProt.R
\name{DATAFRAME_MakeALineForABlastProt}
\alias{DATAFRAME_MakeALineForABlastProt}
\title{Make a dataframe to be use in a future data table that will represent the protein, the peptides that match this protein, the other proteins matched with these peptides.
This function is dedicated to Blast forms, since it will compare two sequences and higlight differences.}
\usage{
DATAFRAME_MakeALineForABlastProt(
  LeadingRazorProtein,
  DT_Peptides,
  Rescue,
  BS_ProteinBank,
  ProteinWidth = 60
)
}
\arguments{
\item{LeadingRazorProtein}{Protein name as displayed in the MaxQuant seach file}

\item{DT_Peptides}{Peptide data table since we are using 'Leading razor protein' and not Leading.razor.protein}

\item{Rescue}{Dataframe of a blast results}

\item{BS_ProteinBank}{AA StringSet protein bank.}

\item{ProteinWidth}{Length of the protein displayed width. expressed in AA}
}
\value{
Dataframe with every elements needed in the displayed data frame
Name name of the protein that will be displayed
BlastHit name of the protein from the known protein bank that match the displayed protein
N_Peptides Number of peptides that match this protein
Match_Proteins Proteins match by this set of peptides
PEPTIDES kable HTML ready to print table of the peptide that match this protein
N_Match_Proteins Number of differents proteins that match this set of peptides
Sequence HTML Ready to print sequence of the protein
SAAV Boolean that indicate if this protein have an SAAV
SAAVxPep Boolean that indicate if this protein have peptide that display an SAAV
Insertion Boolean that indicate if this protein have an Insertion
InsertionxPep Boolean that indicate if this protein have a peptide that displayed a Insertion
Deletion Boolean that indicate if this protein have an deletion
DeletionxPep Boolean that indicate if this protein have a peptide that displayed a Deletion
OutOfBlast Boolean that indicate if this protein have a peptide that display a section out of the blast, i.e. a part that is specific
}
\description{
Make a dataframe to be use in a future data table that will represent the protein, the peptides that match this protein, the other proteins matched with these peptides.
This function is dedicated to Blast forms, since it will compare two sequences and higlight differences.
}
\examples{
library(data.table)
library(Biostrings)
library(DT)
PB_FilePath <- system.file("extdata",
 "protein_bank.example.fasta",
  package = "isoformnspectRe")
ProteinBank<-readAAStringSet(
  PB_FilePath
)
P_FilePath <- system.file("extdata",
 "peptides.example.txt",
  package = "isoformnspectRe")
Peptides<-fread(
 P_FilePath,
 integer64="double")
R_FilePath <- system.file("extdata",
 "blast_rescue.example.txt",
  package = "isoformnspectRe")
Rescue<-read.table(
 R_FilePath,
 header=TRUE)

names(ProteinBank) <- gsub(
 " .*$",
 "",
 names(ProteinBank))
LeadingRazorProtein <-
"strngt.16858.3,start=259,stop=4446,frame=1,flag=LonguestORF,Origin=3',Blast=sp|Q15596|NCOA2_HUMAN"
DF <- DATAFRAME_MakeALineForABlastProt(
LeadingRazorProtein,
Peptides,
Rescue,
ProteinBank)
datatable(DF,escape=FALSE)
ToRepresent<-DF[,c(
 "Name",
 "BlastHit",
 "N_Peptides",
 "Match_Proteins",
 "N_Match_Proteins",
 "Sequence")]
datatable(ToRepresent,escape=FALSE)

}
