% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATAFRAME_MakeALineForAIsoProt.R
\name{DATAFRAME_MakeALineForAIsoProt}
\alias{DATAFRAME_MakeALineForAIsoProt}
\title{Make a dataframe to be use in a future data table that will represent the protein, the peptides that match this protein, the other proteins matched with these peptides.
This function can handle the Canonical form and also the isoforms}
\usage{
DATAFRAME_MakeALineForAIsoProt(
  LeadingRazorProtein,
  DT_Peptides,
  BS_ProteinBank,
  ProteinWidth = 60
)
}
\arguments{
\item{LeadingRazorProtein}{Protein name as displayed in the MaxQuant seach file}

\item{DT_Peptides}{Peptide data table since we are using 'Leading razor protein' and not Leading.razor.protein}

\item{BS_ProteinBank}{AA StringSet protein bank.}

\item{ProteinWidth}{Length of the protein displayed width. expressed in AA}
}
\value{
Dataframe with every elements needed in the displayed data frame
Name name of the protein that will be displayed
N_Peptides Number of peptides that match this protein
PEPTIDES kable HTML ready to print table of the peptide that match this protein
N_Match_Proteins Number of differents proteins that match this set of peptides
Sequence HTML Ready to print sequence of the protein
}
\description{
Make a dataframe to be use in a future data table that will represent the protein, the peptides that match this protein, the other proteins matched with these peptides.
This function can handle the Canonical form and also the isoforms
}
\examples{
library(data.table)
library(Biostrings)
library(DT)
PB_FilePath <- system.file("extdata",
 "protein_bank.example.fasta",
  package = "isoformnspectRe")
ProteinBank<-readAAStringSet(
  PB_FilePath
)
P_FilePath <- system.file("extdata",
 "peptides.example.txt",
  package = "isoformnspectRe")
Peptides<-fread(
 P_FilePath,
 integer64="double")
names(ProteinBank) <- gsub(
 " .*$",
 "",
 names(ProteinBank))
LeadingRazorProtein <-
"strngt.5372.8,start=14,stop=904,frame=2,flag=1,UNIPROT=sp|Q86U42-2|PABP2_HUMAN"
DF<-DATAFRAME_MakeALineForAIsoProt(LeadingRazorProtein,Peptides,ProteinBank)
datatable(DF,escape=FALSE)
}
