% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRangesFromUniprot.R
\name{GetEmptySpanRanges}
\alias{GetEmptySpanRanges}
\title{GetEmptySpanRanges Get an empty IRange}
\usage{
GetEmptySpanRanges()
}
\value{
span an empty Genomic Range with class and title attribute for HTML span inclusion
}
\description{
GetEmptySpanRanges Get an empty IRange
}
\examples{
EmptyRange<-GetEmptySpanRanges()
EmptyRange
}
