% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{HandleFoldChange}
\alias{HandleFoldChange}
\title{Handle fold change
Will substitute 15 to a more extreme value}
\usage{
HandleFoldChange(FoldChanges)
}
\arguments{
\item{FoldChanges}{vector of fold changes (numeric vector)}
}
\value{
vector of numeric values with absolute values set to 15
}
\description{
Handle fold change
Will substitute 15 to a more extreme value
}
\examples{
FC<-c(+Inf,9,3,-20,-Inf)
HandleFoldChange(FC)
}
