% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{HTML_HandleEndOfLine}
\alias{HTML_HandleEndOfLine}
\title{HTML_HandleEndOfLine add a end of a line}
\usage{
HTML_HandleEndOfLine(EndOfTheLine, LineWidth)
}
\arguments{
\item{EndOfTheLine}{At what position you are at the end of this line}

\item{LineWidth}{The length of the line}
}
\value{
charater chain with
}
\description{
HTML_HandleEndOfLine add a end of a line
}
\examples{
HTML_HandleEndOfLine(EndOfTheLine=277,LineWidth=60)
}
