% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{HTML_Pretty_Sequence}
\alias{HTML_Pretty_Sequence}
\title{HTML_Pretty_Sequence Prepare an HTML ready}
\usage{
HTML_Pretty_Sequence(
  Sequence,
  Peptides,
  Rescue = NULL,
  PerfectIsoformName = NULL,
  LineWidth = 60
)
}
\arguments{
\item{Sequence}{String of the protein AA Sequence}

\item{Peptides}{data.frame of the peptides attributed to this protein}

\item{Rescue}{Not Used for now}

\item{PerfectIsoformName}{Name of the Isoform as observed in the}

\item{LineWidth}{Number of AA before a new line (will not include the space inside)}
}
\value{
HTML string of the sequence embedded with features included as span
}
\description{
HTML_Pretty_Sequence Prepare an HTML ready
}
\examples{
library(Biostrings)
PB_FilePath <- system.file("extdata",
 "protein_bank.example.fasta",
  package = "isoformnspectRe")
ProteinBank<-readAAStringSet(
  PB_FilePath
)
Sequence <- as.vector(ProteinBank["Example"])
Peptides <- data.frame(
 Start = c(2, 2, 240),
 End = c(10, 11, 253),
 Pep = c(1.9437e-06, 0.00014334, 0.0001654),
 Proteins = c(
  "strngt.11487.1 Q8WUM4-3 strngt.11487.3 strngt.11487.2",
  "strngt.11487.1 Q8WUM4-3 strngt.11487.3 strngt.11487.2",
  "strngt.11487.2"),
 Sequence = c("ATFISVQLK", "ATFISVQLKK", "YFYFQEVFPVLAAK ")
)
Sequence <- HTML_Pretty_Sequence(Sequence,Peptides)
}
