% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{HTML_VECTOR_Peptide2Popover}
\alias{HTML_VECTOR_Peptide2Popover}
\title{Alter a peptide sequence to have an harmonious presentation}
\usage{
HTML_VECTOR_Peptide2Popover(Sequences, MaxNChar = 15)
}
\arguments{
\item{Sequences}{Vector of peptides Sequence}

\item{MaxNChar}{Maximum width size of the peptide}
}
\value{
string vector as a HTML span to have shorter form
}
\description{
Alter a peptide sequence to have an harmonious presentation
}
\examples{
library("DT")
library("data.table")
P_FilePath <- system.file("extdata",
 "peptides.example.txt",
  package = "isoformnspectRe")
Peptides<-fread(
 P_FilePath,
 integer64="double")
Peptides<-Peptides[,c("Sequence",
"Start position","End position")]
datatable(Peptides)
Peptides$Sequence<-HTML_VECTOR_Peptide2Popover(Peptides$Sequence)
datatable(Peptides,escape=FALSE,
options=list(
 drawCallback=JS(
 paste("$('[data-toggle=\"popover\"]').popover({container: 'body'});")
 )))
}
