% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{LocalOptions}
\alias{LocalOptions}
\title{Title}
\usage{
LocalOptions(
  LIST_SampleNames,
  LIST_SparlinesOptions = list(type = "bar", height = 40, width = 60, highlightColor =
    "red", chartRangeMin = 0, chartRangeMax = 1, tooltipFormat =
    "{{offset:offset}} {{value.2}}")
)
}
\arguments{
\item{LIST_SampleNames}{list of sample names}

\item{LIST_SparlinesOptions}{options for sparkline}
}
\value{
List of options for data table including sparklines
}
\description{
Title
}
\examples{
SamplesNames <- list(
group_un=c("echantillon un","echantillon deux"),
group_deux=c("echantillon trois","echantillon quatre"))
}
