% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{server}
\alias{server}
\title{shiny server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny session}
}
\value{
shiny server
}
\description{
shiny server
}
\examples{
library("isoformnspectRe")
if(interactive()){
GlobalPath<-system.file("extdata",
 "Global.R",
  package = "isoformnspectRe")
source(GlobalPath,local=TRUE)
shiny::shinyApp(ui,server)
}
}
