% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRangesFromUniprot.R
\name{SplitSpanRanges}
\alias{SplitSpanRanges}
\title{SplitSpanRanges Split span ranges by insert Ranges. The span ranges will be break in two if they overlap an insert IRanges. Typically for the endline.}
\usage{
SplitSpanRanges(SpanRanges, InsertRanges)
}
\arguments{
\item{SpanRanges}{Information range to be split if need be}

\item{InsertRanges}{Range of the position to split. Shoudl be of width 1}
}
\value{
SpanRanges with splitted
}
\description{
SplitSpanRanges Split span ranges by insert Ranges. The span ranges will be break in two if they overlap an insert IRanges. Typically for the endline.
}
\examples{
library(GenomicRanges)
VariationRanges<-GetRangesFromUniprot("O94875-10")
Insert<-IRanges(60,60)
SplitSpanRanges(VariationRanges,Insert)
}
