% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRING_ProteinShortName.R
\name{STRING_ProteinShortName}
\alias{STRING_ProteinShortName}
\title{STRING_ProteinShortName Handle the name of a protein as presented in the Maxquant peptide file, to be presented in the kable}
\usage{
STRING_ProteinShortName(ProteinFullName)
}
\arguments{
\item{ProteinFullName}{name as observed in the MaxQuant peptide file}
}
\value{
short name with the UNIPROT code if present. b is for blast equivalent.
}
\description{
STRING_ProteinShortName Handle the name of a protein as presented in the Maxquant peptide file, to be presented in the kable
}
\examples{
LeadingRazorProtein <-
"strngt.5372.8,start=14,stop=904,frame=2,flag=1,UNIPROT=sp|Q86U42-2|PABP2_HUMAN"
STRING_ProteinShortName(LeadingRazorProtein)
}
