% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML.R
\name{UpdateSequence}
\alias{UpdateSequence}
\title{UpdateSequence Transform a AA sequence to a HTML ready to print string}
\usage{
UpdateSequence(IR, PROTEIN_AA)
}
\arguments{
\item{IR}{Integer Ranges with all necessay informations}

\item{PROTEIN_AA}{AA sequence}
}
\value{
a string with span tag
}
\description{
UpdateSequence Transform a AA sequence to a HTML ready to print string
}
