% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIST_INDEX_VECTOR_GetMatchs.R
\name{DATATABLE_AnnotateProteotypic}
\alias{DATATABLE_AnnotateProteotypic}
\title{DATAFRAME_AnnotateProteotypic Since we observe that some proteins are not present in the Proteins field, we decieded to}
\usage{
DATATABLE_AnnotateProteotypic(PeptideDataFrame, ProteinBank, Threads = 1)
}
\arguments{
\item{PeptideDataFrame}{MaxQuant peptide search as a data frame}

\item{ProteinBank}{AAStringSet of the protein fasta file that was used for the mass spectrometry analysis}

\item{Threads}{Number of parallel sessions (numeric)}
}
\value{
PeptideDataFrame Updated data frame with informations
}
\description{
DATAFRAME_AnnotateProteotypic Since we observe that some proteins are not present in the Proteins field, we decieded to
}
